Add search
This adds a search functionality to the website, powered by lunr.js. While data is prepared and collected during build time, the index is built on the client, as the precomputed index is 1.2 MB, which is already the size of a full page load (even if the data is only fetched once you actually search something). I wasn't able to make this work with dynamically routed pages, although it is fully compatible with all our generated pages, as the content is fetched from the built HTML. I also haven't included pagination and mobile support, probably something for a future improvement.