Restore/improve patchnote behaviour (#220)
* Fix version button on mobile

Has the same behaviour as before it broke. With one exception, the sidebar is now automatically opened if the url fragment is empty.
Fixes #219.

* Fix stickyness of sidebar

The overflow-x property was interfering, causing the stickyness to not work. I confirmed on the wayback machine that the stickyness previously did work around March 2022.
Unsetting the overflow-x property did not appear to have any effect on the way overflow is handled.

* Fix scrolling

The buttons used to have the same id as the contents of the url fragment. This caused the browser to scroll to the button. This has a very undesireable and jumpy effect, it also conflicts with our own scrolling.

* Fix sticky scrolling with short changelogs

* Scroll to above the title, not exactly the title

* Add aria roles to mobile menu

---------

Co-authored-by: ix0rai <ix0rai64@gmail.com>