Another new package, woohoo!
Astro doesn’t provide a Markdown processor outside of the content collection loading context, so if you want to render Markdown in any situation other than a custom loader or importing a Markdown file in src/pages, a 3rd-party library is required. Fortunately, this one is designed for this exact scenario.