Upgrade RSS feed to content collections (again)
Switch to Shiki highlighting
We had a custom prism theme but due to a bug with the Markdown parser it’s that or the RSS

Add package for configuring custom Markdown rendering configurations.
Sorry…

Use custom Markdown configuration for RSS posts
It means we can keep Prism syntax highlighting on the site while still having the RSS work.